hotel_ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
anthropic==0.39.0
google-generativeai==0.8.3
groq==0.11.0
httpx==0.27.2
python-json-logger==2.0.7
orjson==3.10.7
tenacity==9.0.0
# IANA timezone data — required on Windows (and minimal containers)
# for `zoneinfo.ZoneInfo("Europe/London")` etc. UTC is handled
# without it. Harmless on platforms that ship tzdata system-wide.
tzdata==2025.2
pytest==8.3.3
pytest-asyncio==0.24.0
